HIVE-18195: Hive schema broken on postgres (Deepesh Khandelwal, reviewed by Sergey Shelukhin)
(cherry picked from commit 4218629de715df43a1778de03f85e41bc682b1a8)

+++ metastore/scripts/upgrade/postgres/045-HIVE-17566.postgres.sql
@@ -17,7 +17,7 @@ CREATE TABLE "WM_POOL" (
     "POOL_ID" bigint NOT NULL,
     "RP_ID" bigint NOT NULL,
     "PATH" character varying(1024) NOT NULL,
-    "ALLOC_FRACTION" DOUBLE,
+    "ALLOC_FRACTION" double precision,
     "QUERY_PARALLELISM" integer,
     "SCHEDULING_POLICY" character varying(1024)
 );
